LazyTensor Cleanup - Move class DeviceContextArena out
