+++ lazy_tensor_core/lazy_tensor_core/csrc/tensor.cpp
@@ -15,6 +15,7 @@
 #include "lazy_tensor_core/csrc/helpers.h"
 #include "lazy_tensor_core/csrc/ir_dump_util.h"
 #include "lazy_tensor_core/csrc/layout_manager.h"
+#include "lazy_tensor_core/csrc/lazy_graph_executor.h"
 #include "lazy_tensor_core/csrc/op_by_op_executor.h"
 #include "lazy_tensor_core/csrc/ops/arithmetic_ir_ops.h"
 #include "lazy_tensor_core/csrc/ops/cast.h"
@@ -203,13 +204,6 @@ DataCacheArena::DataCache* GetDataCache(const Device& device) {
   return arena->Get(device);
 }
 
-ir::Value IrValueFromScalar(const at::Scalar& value, at::ScalarType scalar_type,
-                            const Device& device) {
-  at::Tensor tensor = at::scalar_tensor(value, at::TensorOptions(scalar_type));
-  lazy_tensors::ComputationClient::DataPtr device_data =
-      TensorToDataHandle(tensor, device);
-  return ir::MakeNode<ir::ops::DeviceData>(std::move(device_data));
-}
 
 lazy_tensors::ComputationClient::DataPtr GetDeviceData(const at::Tensor& tensor,
                                                        const Device& device) {
@@ -266,134 +260,6 @@ bool TensorsHaveIR(const std::vector<LazyTensor>& tensors) {
 
 }  // namespace
 
-// The DeviceContextArena holds per device live information and statistics,
-// among which the lazy tensors which are currently alive in the system. This is
-// used to create computation "barriers" in order to flush pending operations
-// and ensure the same computations are created during the training loops.
-class LazyTensor::DeviceContextArena {
-  struct DeviceContext {
-    std::mutex lock;
-    std::map<lazy_tensors::int64, std::weak_ptr<Data>> tensors_data;
-    lazy_tensors::uint64 seed = 101;
-    lazy_tensors::uint64 running_seed = 101;
-    ir::Value seed_ir_value;
-  };
-
- public:
-  static DeviceContextArena* Get() {
-    static DeviceContextArena* arena = new DeviceContextArena();
-    return arena;
-  }
-
-  void RegisterTensor(std::shared_ptr<Data> data) {
-    DeviceContext* devctx = GetDeviceContext(data->device);
-    std::lock_guard<std::mutex> lock(devctx->lock);
-    devctx->tensors_data.emplace(data->unique_id, data);
-    LTC_COUNTER("CreateLtcTensor", 1);
-  }
-
-  void UnregisterTensor(Data* data) {
-    DeviceContext* devctx = GetDeviceContext(data->device);
-    std::lock_guard<std::mutex> lock(devctx->lock);
-    devctx->tensors_data.erase(data->unique_id);
-    LTC_COUNTER("DestroyLtcTensor", 1);
-  }
-
-  std::vector<LazyTensor> GetLiveTensors(const Device* device) {
-    std::vector<LazyTensor> tensors;
-    auto fn = [&](DeviceContext* devctx) {
-      std::lock_guard<std::mutex> lock(devctx->lock);
-      for (auto& uid_wptr : devctx->tensors_data) {
-        std::shared_ptr<Data> data = uid_wptr.second.lock();
-        if (data != nullptr) {
-          tensors.push_back(LazyTensor(std::move(data)));
-        }
-      }
-    };
-    ForAllDeviceContexts(fn, device);
-    return tensors;
-  }
-
-  ir::Value GetRngSeed(const Device& device) {
-    static const at::ScalarType kSeedType = at::ScalarType::Long;
-    static const lazy_tensors::uint64 kSeedMul = 214013;
-    static const lazy_tensors::uint64 kSeedAdd = 2531011;
-    DeviceContext* devctx = GetDeviceContext(device);
-    std::lock_guard<std::mutex> lock(devctx->lock);
-    if (!devctx->seed_ir_value) {
-      devctx->seed_ir_value =
-          IrValueFromScalar(MakeIntScalar(devctx->seed), kSeedType, device);
-    }
-    // Keep the running seed as scalar as well, so we can return it directly
-    // without executing graphs.
-    devctx->running_seed = kSeedAdd + kSeedMul * devctx->running_seed;
-    // Compose new seeds from the root seed, to avoid creating too many
-    // computation parameters which might overflow the device capacity.
-    ir::Value k = ir::ops::ScalarOp(MakeIntScalar(kSeedMul),
-                                    MakeLtcPrimitiveType(kSeedType, &device));
-    ir::Value b = ir::ops::ScalarOp(MakeIntScalar(kSeedAdd),
-                                    MakeLtcPrimitiveType(kSeedType, &device));
-    devctx->seed_ir_value = b + k * devctx->seed_ir_value;
-    return devctx->seed_ir_value;
-  }
-
-  lazy_tensors::uint64 GetRunningSeed(const Device& device) {
-    DeviceContext* devctx = GetDeviceContext(device);
-    std::lock_guard<std::mutex> lock(devctx->lock);
-    return devctx->running_seed;
-  }
-
-  void SetRngSeed(const Device& device, lazy_tensors::uint64 seed) {
-    DeviceContext* devctx = GetDeviceContext(device);
-    std::lock_guard<std::mutex> lock(devctx->lock);
-    devctx->seed = seed;
-    devctx->running_seed = devctx->seed;
-    devctx->seed_ir_value = ir::Value();
-  }
-
-  void MarkStep(const Device& device) {
-    DeviceContext* devctx = GetDeviceContext(device);
-    std::lock_guard<std::mutex> lock(devctx->lock);
-    devctx->seed = 1012031 + devctx->seed * 7012063;
-    devctx->running_seed = devctx->seed;
-    devctx->seed_ir_value = ir::Value();
-  }
-
- private:
-  std::vector<DeviceContext*> GetAllDeviceContexts() {
-    std::vector<DeviceContext*> all_device_contexts;
-    std::lock_guard<std::mutex> lock(lock_);
-    all_device_contexts.reserve(device_contexts_.size());
-    for (auto& device_contexts : device_contexts_) {
-      all_device_contexts.push_back(device_contexts.second);
-    }
-    return all_device_contexts;
-  }
-
-  void ForAllDeviceContexts(const std::function<void(DeviceContext*)>& fn,
-                            const Device* device) {
-    if (device == nullptr) {
-      for (auto devctx : GetAllDeviceContexts()) {
-        fn(devctx);
-      }
-    } else {
-      fn(GetDeviceContext(*device));
-    }
-  }
-
-  DeviceContext* GetDeviceContext(const Device& device) {
-    std::lock_guard<std::mutex> lock(lock_);
-    auto it = device_contexts_.find(device);
-    if (it == device_contexts_.end()) {
-      it = device_contexts_.emplace(device, new DeviceContext()).first;
-    }
-    return it->second;
-  }
-
-  std::mutex lock_;
-  std::map<Device, DeviceContext*> device_contexts_;
-};
-
 struct DeviceDataInfo : public lazy_tensors::client::Data::Info {
   DeviceDataInfo(lazy_tensors::int64 tensor_id, bool read_only)
       : tensor_id(tensor_id), read_only(read_only) {}
@@ -402,7 +268,7 @@ struct DeviceDataInfo : public lazy_tensors::client::Data::Info {
   bool read_only = false;
 };
 
-LazyTensor::Data::~Data() { DeviceContextArena::Get()->UnregisterTensor(this); }
+LazyTensor::Data::~Data() { LazyGraphExecutor::Get()->UnregisterTensor(this); }
 
 LazyTensor::Async::Async(
     SyncTensorCollection* coll,
@@ -442,7 +308,7 @@ void LazyTensor::Async::Wait() {
 LazyTensor LazyTensor::Create(const at::Tensor& tensor, const Device& device) {
   LTC_CHECK_NE(tensor.device().type(), at::kLazy);
   LazyTensor xtensor(tensor, device);
-  DeviceContextArena::Get()->RegisterTensor(xtensor.data_ptr());
+  LazyGraphExecutor::Get()->RegisterTensor(xtensor.data_ptr());
   return xtensor;
 }
 
@@ -450,7 +316,7 @@ LazyTensor LazyTensor::Create(
     lazy_tensors::ComputationClient::DataPtr handle,
     c10::optional<at::ScalarType> logical_element_type) {
   LazyTensor xtensor(std::move(handle), logical_element_type);
-  DeviceContextArena::Get()->RegisterTensor(xtensor.data_ptr());
+  LazyGraphExecutor::Get()->RegisterTensor(xtensor.data_ptr());
   return xtensor;
 }
 
@@ -458,7 +324,7 @@ LazyTensor LazyTensor::Create(
     ir::Value ir_value, const Device& device,
     c10::optional<at::ScalarType> logical_element_type) {
   LazyTensor xtensor(std::move(ir_value), device, logical_element_type);
-  DeviceContextArena::Get()->RegisterTensor(xtensor.data_ptr());
+  LazyGraphExecutor::Get()->RegisterTensor(xtensor.data_ptr());
   return xtensor;
 }
 
@@ -466,8 +332,12 @@ LazyTensor LazyTensor::Create(
     std::shared_ptr<View> view, const Device& device,
     c10::optional<at::ScalarType> logical_element_type) {
   LazyTensor xtensor(std::move(view), device, logical_element_type);
-  DeviceContextArena::Get()->RegisterTensor(xtensor.data_ptr());
+  LazyGraphExecutor::Get()->RegisterTensor(xtensor.data_ptr());
   return xtensor;
+}
+
+LazyTensor LazyTensor::Create(std::shared_ptr<Data> data) {
+  return LazyTensor(std::move(data));
 }
 
 LazyTensor::LazyTensor(const at::Tensor& tensor, const Device& device)
@@ -926,7 +796,7 @@ void LazyTensor::UpdateFromTensorOut(const LazyTensor& tensor) {
 }
 
 std::vector<LazyTensor> LazyTensor::GetLiveTensors(const Device* device) {
-  return DeviceContextArena::Get()->GetLiveTensors(device);
+  return LazyGraphExecutor::Get()->GetLiveTensors(device);
 }
 
 std::vector<lazy_tensors::ComputationClient::DataPtr>
@@ -1417,7 +1287,7 @@ void LazyTensor::SyncLiveTensorsGraph(
 
 void LazyTensor::MarkStep(const Device& device) {
   LTC_COUNTER("MarkStep", 1);
-  DeviceContextArena::Get()->MarkStep(device);
+  LazyGraphExecutor::Get()->MarkStep(device);
   ir::ScopePusher::ResetScopes();
   g_tls_data.Reset();
 }
@@ -1649,18 +1519,6 @@ lazy_tensors::int64 LazyTensor::GetNextTensorId() {
   static std::atomic<lazy_tensors::int64>* id_generator =
       new std::atomic<lazy_tensors::int64>(1);
   return id_generator->fetch_add(1);
-}
-
-ir::Value LazyTensor::GetRngSeed(const Device& device) {
-  return DeviceContextArena::Get()->GetRngSeed(device);
-}
-
-void LazyTensor::SetRngSeed(const Device& device, lazy_tensors::uint64 seed) {
-  DeviceContextArena::Get()->SetRngSeed(device, seed);
-}
-
-lazy_tensors::uint64 LazyTensor::GetRunningSeed(const Device& device) {
-  return DeviceContextArena::Get()->GetRunningSeed(device);
 }
 
 }  // namespace torch_lazy_tensors
