+++ lazy_tensor_core/test/cpp/torch_ltc_ts_test.cpp
@@ -2,6 +2,7 @@
 
 #include <ATen/ATen.h>
 
+#include "lazy_tensor_core/csrc/lazy_graph_executor.h"
 #include "lazy_tensor_core/csrc/tensor.h"
 
 namespace torch_lazy_tensors {
@@ -9,7 +10,7 @@ namespace cpp_test {
 
 void LtcTsTest::SetUp() {
   at::manual_seed(42);
-  LazyTensor::SetRngSeed(GetCurrentDevice(), 42);
+  LazyGraphExecutor::Get()->SetRngSeed(GetCurrentDevice(), 42);
 }
 
 void LtcTsTest::TearDown() {}
